requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
appdirs==1.4.4
attrs==25.4.0
beautifulsoup4==4.14.3
bert-score==0.3.13
certifi==2026.1.4
cfgv==3.5.0
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
dataclasses-json==0.6.7
datasets==4.4.2
dill==0.4.0
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
docstring_parser==0.17.0
filelock==3.20.2
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.10.0
google-auth==2.47.0
google-genai==1.57.0
groq==0.37.1
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
identify==2.6.15
idna==3.11
instructor==1.13.0
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.9
langchain==1.2.0
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.6
langchain-groq==1.1.1
langchain-huggingface==1.2.0
langchain-openai==1.1.6
langchain-text-splitters==1.1.0
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.1
langsmith==0.6.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.6.1
nodeenv==1.10.0
numpy==2.4.0
openai==2.14.0
orjson==3.11.5
ormsgpack==1.12.1
packaging==25.0
pandas==2.3.3
pillow==12.1.0
platformdirs==4.5.1
pre_commit==4.5.1
propcache==0.4.1
protobuf==6.33.3
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
ragas==0.4.2
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scikit-network==0.33.5
scipy==1.16.3
seaborn==0.13.2
sentence-transformers==5.2.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.1
SQLAlchemy==2.0.45
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
torch==2.9.1
tqdm==4.67.1
transformers==4.57.3
ty==0.0.8
typer==0.21.0
typer-slim==0.21.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
uuid_utils==0.12.0
virtualenv==20.35.4
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
